% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_graph.R
\name{CG_discrete_mutual_information}
\alias{CG_discrete_mutual_information}
\title{Correlation graph via Mutual information}
\usage{
CG_discrete_mutual_information(mat, threshold = 0.7, edge_count = NA,
  numBins = 5, ...)
}
\arguments{
\item{mat}{data}

\item{threshold}{threshold}

\item{edge_count}{number of edges}

\item{numBins}{discretization level}

\item{...}{not used}
}
\value{
adjacency matrix
}
\description{
Correlation graph via Mutual information
}
