% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_graph.R
\name{CG_kendalls_tau}
\alias{CG_kendalls_tau}
\title{Correlation graph via Kendall's tau}
\usage{
CG_kendalls_tau(mat, threshold = 0.7, edge_count = NA, ...)
}
\arguments{
\item{mat}{data}

\item{threshold}{threshold}

\item{edge_count}{number of edges}

\item{...}{not used}
}
\value{
adjacency matrix
}
\description{
Correlation graph via Kendall's tau
}
