% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{contribution}
\alias{contribution}
\title{Contribution object}
\usage{
contribution(breakpoints, values, samples = NA)
}
\arguments{
\item{breakpoints}{vector of numerics (breakpoints, x-direction), must be sorted}

\item{values}{vector of numerics (downstream value)}

\item{samples}{(optional) 3-column numeric matrix to store in the contribution
matrix, used solely for plotting purposes.}
}
\description{
Each of these functions are right continuous.
}
