% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example dataset}
\format{A list with a matrix \code{mat} and data frame \code{pheno}. There
are 107 samples.
\describe{
  \item{mat}{matrix with 2 columns}
  \item{pheno}{data frame with 1 column, \code{Days}}
}}
\description{
Example dataset
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}
