% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution_evaluate.R
\name{evaluate.contribution}
\alias{evaluate.contribution}
\title{Evaluates contribution output}
\usage{
\method{evaluate}{contribution}(obj, target, ...)
}
\arguments{
\item{obj}{contribution object to evaluate}

\item{target}{numeric (scalar) to evaluate}

\item{...}{not used}
}
\value{
numeric
}
\description{
Evaluates contribution output
}
