% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{evaluate}
\alias{evaluate}
\title{Evaluates an object}
\usage{
evaluate(obj, target, ...)
}
\arguments{
\item{obj}{object to evaluate}

\item{target}{input to evaluate the obj at}

\item{...}{not used}
}
\value{
vector
}
\description{
Evaluates an object
}
