% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cognition_evaluate.R
\name{evaluate.stat_cognition}
\alias{evaluate.stat_cognition}
\title{Evaluates stat cognition output}
\usage{
\method{evaluate}{stat_cognition}(obj, target, ...)
}
\arguments{
\item{obj}{stat_cognition object to evaluate}

\item{target}{input data object to evaluate}

\item{...}{not used}
}
\value{
stat_cognition_result object
}
\description{
Evaluates stat cognition output
}
