% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_evaluate.R
\name{evaluate.value}
\alias{evaluate.value}
\title{Evaluates value output}
\usage{
\method{evaluate}{value}(obj, target, return.value = T, ...)
}
\arguments{
\item{obj}{value object to evaluate}

\item{target}{input vector (evaluted from the state functions) to evaluate}

\item{return.value}{boolean. If TRUE, return the numeric value of the value
function. If FALSE, return the integer corresponding to which action was taken}

\item{...}{not used}
}
\value{
numeric
}
\description{
Evaluates value output
}
