% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_decoupling.R
\name{generator_add_noise}
\alias{generator_add_noise}
\title{Synthetic generator function: Add Gaussian noise to columns}
\usage{
generator_add_noise(dat, param1 = c(0, 1), param2 = c(0, 1),
  param3 = c(-0.1, 0.1), param4 = c(1e-04, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of columns}

\item{param2}{parameter for percentage of rows}

\item{param3}{parameter for mean of noise}

\item{param4}{parameter for sd of noise}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Add Gaussian noise to columns
}
