% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_coupling.R
\name{generator_brownian}
\alias{generator_brownian}
\title{Synthetic generator function: Adjust a pair of variables based on Brownian motion}
\usage{
generator_brownian(dat, param1 = c(0, 0.5), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for shrinkage}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Adjust a pair of variables based on Brownian motion
}
