% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_coupling.R
\name{generator_circle}
\alias{generator_circle}
\title{Synthetic generator function: Shrink points around a circle for a pair of points}
\usage{
generator_circle(dat, param1 = c(0, 1), param2 = c(0, 1), param3 = c(0,
  1), param4 = c(0, 0.5), param5 = c(0, 6.28), param6 = c(0, 6.28),
  param7 = c(0, 1.57), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of rows}

\item{param2}{parameter for percentage width of circle}

\item{param3}{parameter for percentage height of circle}

\item{param4}{parameter for shrinkage}

\item{param5}{parameter for start of radians}

\item{param6}{parameter for number of radians}

\item{param7}{parameter for angle of circle}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Shrink points around a circle for a pair of points
}
