% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions.R
\name{generator_cluster}
\alias{generator_cluster}
\title{Synthetic generator function: Cluster points}
\usage{
generator_cluster(dat, param1 = c(0, 1), param2 = c(0, 1), param3 = c(0,
  10), param4 = c(-1, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of columns}

\item{param2}{parameter for percentage of rows}

\item{param3}{parameter for number of clusters}

\item{param4}{parameter for shrinkage}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Cluster points
}
