% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_pheno.R
\name{generator_cluster_pheno}
\alias{generator_cluster_pheno}
\title{Synthetic generator function: Cluster phenotype and variable}
\usage{
generator_cluster_pheno(dat, param1 = c(0, 1), param2 = c(0, 1),
  param3 = c(2, 10), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of rows}

\item{param2}{parameter for percentage of columns}

\item{param3}{parameter for number of clusters}

\item{...}{not used}
}
\value{
data object
}
\description{
Pick one variable in mat, and one continuous variable in pheno
}
