% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_decoupling.R
\name{generator_decouple_empirical}
\alias{generator_decouple_empirical}
\title{Synthetic generator function: Make data independent}
\usage{
generator_decouple_empirical(dat, param1 = c(0, 1), param2 = c(0, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of columns}

\item{param2}{parameter for percentage of rows}

\item{...}{not used}
}
\value{
data object
}
\description{
Sample from the empirical distribution and add some noise
}
