% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions.R
\name{generator_inflate_correlation}
\alias{generator_inflate_correlation}
\title{Synthetic generator function: Inflate the correlation between pairs of variables}
\usage{
generator_inflate_correlation(dat, param1 = c(0, 10), param2 = c(0, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for number of pairs to consider}

\item{param2}{parameter for percentage of rows}

\item{...}{not used}
}
\value{
data object
}
\description{
Set one variable to the mean of itself and another variable
}
