% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions.R
\name{generator_monotonic}
\alias{generator_monotonic}
\title{Synthetic generator function: Make a pair of variables monotonically dependent}
\usage{
generator_monotonic(dat, param1 = c(0, 1), param2 = c(0, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of rows}

\item{param2}{parameter for strength of monotonicity}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Make a pair of variables monotonically dependent
}
