% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_pheno.R
\name{generator_monotonic_pheno}
\alias{generator_monotonic_pheno}
\title{Synthetic generator function: Sort pheno in order with respect to one variable}
\usage{
generator_monotonic_pheno(dat, param1 = c(0, 1), param2 = c(0, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of rows to refit}

\item{param2}{parameter for strength of monotonicity}

\item{...}{not used}
}
\value{
data object
}
\description{
Pick one variable in mat, and one continuous variable in pheno
}
