% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_coupling.R
\name{generator_nearest_neighbor}
\alias{generator_nearest_neighbor}
\title{Synthetic generator function: Shrink towards nearest neighbor}
\usage{
generator_nearest_neighbor(dat, param1 = c(0, 1), param2 = c(0, 0.5), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of rows}

\item{param2}{parameter for shrinkage}

\item{...}{not used}
}
\value{
data object
}
\description{
Shrink based on all dimensions
}
