% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_decoupling.R
\name{generator_outlier}
\alias{generator_outlier}
\title{Synthetic generator function: Create outliers}
\usage{
generator_outlier(dat, param1 = c(1, 5), param2 = c(1, 3), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for number of points}

\item{param2}{parameter for expansion}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Create outliers
}
