% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_coupling.R
\name{generator_polynomial}
\alias{generator_polynomial}
\title{Synthetic generator function: Polynomial regression on a pair of variables}
\usage{
generator_polynomial(dat, param1 = c(-1, 1), param2 = c(3, 6),
  param3 = c(0, 0.5), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of rows}

\item{param2}{parameter for degree of polynomial}

\item{param3}{parameter for shrinkage}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Polynomial regression on a pair of variables
}
