% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions.R
\name{generator_refit_normality}
\alias{generator_refit_normality}
\title{Synthetic generator function: Refit from Gaussian distribution}
\usage{
generator_refit_normality(dat, param1 = c(0, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of rows to refit}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Refit from Gaussian distribution
}
