% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions_decoupling.R
\name{generator_resample}
\alias{generator_resample}
\title{Synthetic generator function: Resample rows}
\usage{
generator_resample(dat, param1 = c(0, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of rows to refit}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Resample rows
}
