% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_functions.R
\name{generator_shuffle}
\alias{generator_shuffle}
\title{Synthetic generator function: Shuffle data}
\usage{
generator_shuffle(dat, param1 = c(0, 1), param2 = c(0, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of columns}

\item{param2}{parameter for percentage of rows}

\item{...}{not used}
}
\value{
data object
}
\description{
Synthetic generator function: Shuffle data
}
