% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_object.R
\name{is_valid.data}
\alias{is_valid.data}
\title{Checks data object for validity}
\usage{
\method{is_valid}{data}(obj, ...)
}
\arguments{
\item{obj}{The object to check}

\item{...}{not used}
}
\value{
boolean
}
\description{
Runs the functions on the minimum and maximum values on the dataset
}
