% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{is_valid}
\alias{is_valid}
\title{Checks object validity}
\usage{
is_valid(obj, ...)
}
\arguments{
\item{obj}{object to check}

\item{...}{not used}
}
\value{
boolean
}
\description{
Generic function that fails noisily with a stop message if the object is
invalid. Otherwise, returns TRUE
}
