% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_contribution.R
\name{is_valid.state_action}
\alias{is_valid.state_action}
\title{Checks state_action object for validity}
\usage{
\method{is_valid}{state_action}(obj, ...)
}
\arguments{
\item{obj}{The object to check}

\item{...}{not used}
}
\value{
boolean
}
\description{
Checks state_action object for validity
}
