% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_initializer.R
\name{is_valid.synthetic_initializer}
\alias{is_valid.synthetic_initializer}
\title{Checks synthetic_initializer object for validity}
\usage{
\method{is_valid}{synthetic_initializer}(obj, dat, ...)
}
\arguments{
\item{obj}{The object to check}

\item{dat}{data object to test all the synthetic generator functions on}

\item{...}{not used}
}
\value{
boolean
}
\description{
Runs the functions on the minimum and maximum values on the dataset
}
