% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{is_valid.value}
\alias{is_valid.value}
\title{Checks value object for validity}
\usage{
\method{is_valid}{value}(obj)
}
\arguments{
\item{obj}{The object to check}
}
\value{
boolean
}
\description{
Checks value object for validity
}
