% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_helpers.R
\name{mixing_inflate_correlation}
\alias{mixing_inflate_correlation}
\title{Mixing function: Inflate correlation}
\usage{
mixing_inflate_correlation(mat, percentage_columns = 0.1, ...)
}
\arguments{
\item{mat}{data matrix}

\item{percentage_columns}{percentage of columns affected}

\item{...}{not used}
}
\value{
numeric matrix
}
\description{
Mixing function: Inflate correlation
}
