% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_helpers.R
\name{mixing_refit_normality}
\alias{mixing_refit_normality}
\title{Mixing function: Refit from Gaussian distribution}
\usage{
mixing_refit_normality(dat, param1 = c(0, 1), ...)
}
\arguments{
\item{dat}{data object}

\item{param1}{parameter for percentage of rows to refit}

\item{...}{not used}
}
\value{
data object
}
\description{
Mixing function: Refit from Gaussian distribution
}
