% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_helpers.R
\name{mixing_resample}
\alias{mixing_resample}
\title{Mixing function: Resample rows}
\usage{
mixing_resample(mat, ...)
}
\arguments{
\item{mat}{data matrix}

\item{...}{not used}
}
\value{
numeric matrix
}
\description{
Mixing function: Resample rows
}
