% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_helpers.R
\name{mixing_shuffle_columns}
\alias{mixing_shuffle_columns}
\title{Mixing function: Shuffle columns}
\usage{
mixing_shuffle_columns(mat, percentage_columns = 0.1, ...)
}
\arguments{
\item{mat}{data matrix}

\item{percentage_columns}{percentage of columns affected}

\item{...}{not used}
}
\value{
numeric matrix
}
\description{
Mixing function: Shuffle columns
}
