% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_helpers.R
\name{mixing_swap_rows}
\alias{mixing_swap_rows}
\title{Mixing function: Swap rows}
\usage{
mixing_swap_rows(mat, idx1 = c(1, 2), idx2 = c(3, 4), ...)
}
\arguments{
\item{mat}{data matrix}

\item{idx1}{indicies of rows}

\item{idx2}{indicies of rows}

\item{...}{not used}
}
\value{
numeric matrix
}
\description{
Mixing function: Swap rows
}
