% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value.R
\name{MV_matching}
\alias{MV_matching}
\title{Missing value treatment via matching}
\usage{
MV_matching(mat, ...)
}
\arguments{
\item{mat}{matrix}

\item{...}{not used}
}
\value{
matrix
}
\description{
Missing value treatment via matching
}
