% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value.R
\name{MV_matrix_completion}
\alias{MV_matrix_completion}
\title{Missing value treatment via matrix completion}
\usage{
MV_matrix_completion(mat, pheno = NA, rank = ceiling(ncol(mat)/2),
  lambda = 1, ...)
}
\arguments{
\item{mat}{matrix}

\item{pheno}{data frame}

\item{rank}{rank of the latent factors}

\item{lambda}{tuning parameter}

\item{...}{not used}
}
\value{
matrix
}
\description{
Uses softimpute
}
