% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value.R
\name{MV_remove}
\alias{MV_remove}
\title{Missing value treatment via removal}
\usage{
MV_remove(mat, ...)
}
\arguments{
\item{mat}{matrix}

\item{...}{not used}
}
\value{
matrix
}
\description{
Missing value treatment via removal
}
