% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_dependency.R
\name{PD_kendall}
\alias{PD_kendall}
\title{Pairwise depedency: Kendall's tau}
\usage{
PD_kendall(dat, threshold = 0.35)
}
\arguments{
\item{dat}{data object with only 2 columns}

\item{threshold}{threshold}
}
\value{
boolean
}
\description{
Pairwise depedency: Kendall's tau
}
