% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_dependency.R
\name{PD_pearson}
\alias{PD_pearson}
\title{Pairwise depedency: Pearson}
\usage{
PD_pearson(dat, threshold = 0.35)
}
\arguments{
\item{dat}{data object with only 2 columns}

\item{threshold}{threshold}
}
\value{
boolean
}
\description{
Pairwise depedency: Pearson
}
