% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.contribution.R
\name{plot.contribution}
\alias{plot.contribution}
\title{Plotting function for contribution}
\usage{
\method{plot}{contribution}(x, jitter = T, samples = F, contribution = F,
  lwd = 1, line_lwd = 3, ...)
}
\arguments{
\item{x}{contribution object}

\item{jitter}{boolean for jittering x-axis}

\item{samples}{boolean for plotting samples}

\item{contribution}{boolean splitting samples into total value and contribution}

\item{lwd}{plotting \code{lwd} parameter for samples}

\item{line_lwd}{plotting \code{lwd} parameter for fit}

\item{...}{other plotting parameters}
}
\value{
void
}
\description{
Plotting function for contribution
}
