% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.value.R
\name{plot.value}
\alias{plot.value}
\title{Plotting function for value}
\usage{
\method{plot}{value}(x, type = "contribution", ...)
}
\arguments{
\item{x}{value object}

\item{type}{string for plot type}

\item{...}{additional plotting parameters}
}
\value{
void
}
\description{
Plotting function for value
}
