% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_confounding.R
\name{RC_linear_regression}
\alias{RC_linear_regression}
\title{Removing confounders via linear regression}
\usage{
RC_linear_regression(mat, pheno, ...)
}
\arguments{
\item{mat}{matrix}

\item{pheno}{data frame}

\item{...}{not used}
}
\value{
matrix
}
\description{
Removing confounders via linear regression
}
