% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_confounding.R
\name{RC_none}
\alias{RC_none}
\title{Removing confounders via none}
\usage{
RC_none(mat, pheno, ...)
}
\arguments{
\item{mat}{matrix}

\item{pheno}{data frame}

\item{...}{not used}
}
\value{
matrix
}
\description{
Removing confounders via none
}
