% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_confounding.R
\name{RC_pairing_difference}
\alias{RC_pairing_difference}
\title{Removing confounders via paired difference}
\usage{
RC_pairing_difference(mat, pheno, ...)
}
\arguments{
\item{mat}{matrix}

\item{pheno}{data frame}

\item{...}{not used}
}
\value{
matrix
}
\description{
Removing confounders via paired difference
}
