% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_confounding.R
\name{RC_random_forest_regression}
\alias{RC_random_forest_regression}
\title{Removing confounders via random forest}
\usage{
RC_random_forest_regression(mat, pheno, ...)
}
\arguments{
\item{mat}{matrix}

\item{pheno}{data frame}

\item{...}{not used}
}
\value{
matrix
}
\description{
Removing confounders via random forest
}
