% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_selection.R
\name{SS_neighborhood}
\alias{SS_neighborhood}
\title{Sample selection via neighborhood}
\usage{
SS_neighborhood(dat, quantile = 0.5, neighbor_threshold = 1, ...)
}
\arguments{
\item{dat}{data object}

\item{quantile}{quantile of all the pairwise distances to be the threshold}

\item{neighbor_threshold}{how many neighbors or fewer needed to be considered outlier}

\item{...}{not used}
}
\value{
list containing the modified \code{mat} and \code{pheno}
}
\description{
Sample selection via neighborhood
}
