% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_selection.R
\name{SS_none}
\alias{SS_none}
\title{Sample selection via all}
\usage{
SS_none(mat, ...)
}
\arguments{
\item{mat}{matrix}

\item{...}{not used}
}
\value{
matrix
}
\description{
Sample selection via all
}
