% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_selection.R
\name{SS_quantile}
\alias{SS_quantile}
\title{Sample selection via min/max removal}
\usage{
SS_quantile(mat, ...)
}
\arguments{
\item{mat}{matrix}

\item{...}{not used}
}
\value{
matrix
}
\description{
Sample selection via min/max removal
}
