% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_selection.R
\name{SS_three_sd}
\alias{SS_three_sd}
\title{Sample selection via three standard deviations}
\usage{
SS_three_sd(mat, ...)
}
\arguments{
\item{mat}{matrix}

\item{...}{not used}
}
\value{
matrix
}
\description{
Sample selection via three standard deviations
}
