% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cognition.R
\name{stat_cognition}
\alias{stat_cognition}
\title{Statistical cognition system}
\usage{
stat_cognition(dat, init, seed_vec, response_vec = NA, verbose = F,
  store = F)
}
\arguments{
\item{dat}{data object}

\item{init}{synthetic_initializer object}

\item{seed_vec}{vector of seeds for synthetic data}

\item{response_vec}{optional responses. If not supplied, the system is interactive}

\item{verbose}{boolean on print statements}

\item{store}{boolean to output the contribution functions themselves for visualization
purposes only}
}
\description{
Statistical cognition system
}
