% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cognition_initializer.R
\name{stat_cognition_initializer}
\alias{stat_cognition_initializer}
\title{Statistical cognition system initializer}
\usage{
stat_cognition_initializer(action_ll, state_ll, difference_list = NA,
  generator_init = synthetic_initializer())
}
\arguments{
\item{action_ll}{list of list of actions}

\item{state_ll}{list of list of states}

\item{difference_list}{list of difference functions}

\item{generator_init}{synthetic_initializer object}
}
\value{
stat_cognition_initializer object
}
\description{
Statistical cognition system initializer
}
