% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_data_sample_selection.R
\name{state_data_influential_points}
\alias{state_data_influential_points}
\title{State feature: outliers based on influential points}
\usage{
state_data_influential_points(mat, num_pairs = 50, ...)
}
\arguments{
\item{mat}{data matrix}

\item{num_pairs}{number of pairs of variables}

\item{...}{not used}
}
\value{
value
}
\description{
State feature: outliers based on influential points
}
