% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_data_correlation_graph.R
\name{state_data_monotonicity_pairs}
\alias{state_data_monotonicity_pairs}
\title{State feature: monotonic pairs}
\usage{
state_data_monotonicity_pairs(mat, num_pairs = 50, num_breakpoints = 5, ...)
}
\arguments{
\item{mat}{data matrix}

\item{num_pairs}{number of pairs of variables}

\item{num_breakpoints}{number of breakpoints to estimate}

\item{...}{not used}
}
\value{
value
}
\description{
State feature: monotonic pairs
}
