% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_data_sample_selection.R
\name{state_data_nearest_neighbor}
\alias{state_data_nearest_neighbor}
\title{State feature: outliers based on nearest neighbor distance}
\usage{
state_data_nearest_neighbor(mat, ...)
}
\arguments{
\item{mat}{data matrix}

\item{...}{not used}
}
\value{
value
}
\description{
State feature: outliers based on nearest neighbor distance
}
