% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_data_remove_confounding.R
\name{state_data_pheno_MI}
\alias{state_data_pheno_MI}
\title{State feature: Phenotype mutual information}
\usage{
state_data_pheno_MI(mat, pheno, numBins = 4, ...)
}
\arguments{
\item{mat}{data matrix}

\item{pheno}{data frame phenotype}

\item{numBins}{number of discretized bins}

\item{...}{not used}
}
\value{
value
}
\description{
State feature: Phenotype mutual information
}
