% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_data_remove_confounding.R
\name{state_data_pheno_residual_RF_LR}
\alias{state_data_pheno_residual_RF_LR}
\title{State feature: Phenotype residuals}
\usage{
state_data_pheno_residual_RF_LR(mat, pheno, test_prop = 0.1, quant = 0.75,
  ...)
}
\arguments{
\item{mat}{data matrix}

\item{pheno}{data frame phenotypes}

\item{test_prop}{proportion of values left as test sample}

\item{quant}{quantile of the difference in residuals}

\item{...}{not used}
}
\value{
value
}
\description{
State feature: Phenotype residuals
}
