% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_data_missing_value.R
\name{state_data_residual_RF_LR}
\alias{state_data_residual_RF_LR}
\title{State feature: residuals between random forest and linear regression}
\usage{
state_data_residual_RF_LR(mat, test_prop = 0.1, quant = 0.75, ...)
}
\arguments{
\item{mat}{data matrix}

\item{test_prop}{proportion of values left as test sample}

\item{quant}{quantile of the difference in residuals}

\item{...}{not used}
}
\value{
value
}
\description{
Does prediction of all d variables
}
