% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_sample_selection.R
\name{state_interpoint}
\alias{state_interpoint}
\title{State feature: interpoint distance}
\usage{
state_interpoint(dat, ...)
}
\arguments{
\item{dat}{data object}

\item{...}{not used}
}
\value{
value
}
\description{
Computes the ratio between the maximum (minimum distance to nearest point)
over minimum (minimum distance to nearest point)
}
