% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_correlation.R
\name{state_linearity}
\alias{state_linearity}
\title{State feature: Pairwise linearity}
\usage{
state_linearity(dat, test_prop = 0.1, quant = 0.75, num_pairs = 50, ...)
}
\arguments{
\item{dat}{data object}

\item{test_prop}{proportion of values left as test sample}

\item{quant}{quantile of the difference in residuals}

\item{num_pairs}{maximum number of pairs to look at}

\item{...}{not used}
}
\value{
value
}
\description{
State feature: Pairwise linearity
}
