% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_correlation.R
\name{state_monotone}
\alias{state_monotone}
\title{State feature: monotonic pairs}
\usage{
state_monotone(dat, num_pairs = 5, ...)
}
\arguments{
\item{dat}{data object}

\item{num_pairs}{maximum number of pairs to look at}

\item{...}{not used}
}
\value{
value
}
\description{
For each pair of variables, count the median number of times the
sign changes when using fused lasso
}
