% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_pheno.R
\name{state_pheno_MI}
\alias{state_pheno_MI}
\title{State feature: Phenotype mutual information}
\usage{
state_pheno_MI(dat, numBins = 4, ...)
}
\arguments{
\item{dat}{data object}

\item{numBins}{number of discretized bins}

\item{...}{not used}
}
\value{
value
}
\description{
State feature: Phenotype mutual information
}
