% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_pheno.R
\name{state_pheno_residual}
\alias{state_pheno_residual}
\title{State feature: Phenotype residuals}
\usage{
state_pheno_residual(dat, test_prop = 0.1, quant = 0.75, ...)
}
\arguments{
\item{dat}{data object}

\item{test_prop}{proportion of values left as test sample}

\item{quant}{quantile of the difference in residuals}

\item{...}{not used}
}
\value{
value
}
\description{
State feature: Phenotype residuals
}
