% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_correlation.R
\name{state_samples}
\alias{state_samples}
\title{State feature: proportion of samples remaining}
\usage{
state_samples(dat, prev_dat, ...)
}
\arguments{
\item{dat}{data object}

\item{prev_dat}{previous data object}

\item{...}{not used}
}
\value{
value
}
\description{
State feature: proportion of samples remaining
}
