% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_sample_selection.R
\name{state_variance}
\alias{state_variance}
\title{State feature: variance along PC's}
\usage{
state_variance(dat, num_pc = 1, ...)
}
\arguments{
\item{dat}{data object}

\item{num_pc}{number of principal components to consider}

\item{...}{not used}
}
\description{
State feature: variance along PC's
}
