% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_generator.R
\name{synthetic_generator}
\alias{synthetic_generator}
\title{Generate synthetic data}
\usage{
synthetic_generator(dat, init)
}
\arguments{
\item{dat}{data object}

\item{init}{synthetic_intializer object}
}
\value{
data object with a new field \code{seed}
}
\description{
Generate synthetic data
}
