% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_initializer.R
\name{synthetic_initializer}
\alias{synthetic_initializer}
\title{Synthetic generator initializer}
\usage{
synthetic_initializer(func_list = .grab_package_contents("generator"),
  lambda = 5)
}
\arguments{
\item{func_list}{list of functions that take in \code{data} object and return
another \code{data} object}

\item{lambda}{tuning parameter}
}
\value{
synthetic_intializer object
}
\description{
The number of synthetic functions applied onto the dataset is controlled
by a Poisson random variable with parameter \code{lambda}
}
