% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{value}
\alias{value}
\title{Value function}
\usage{
value(surface)
}
\arguments{
\item{surface}{list of \code{hash} and \code{table}}
}
\value{
value object
}
\description{
\code{surface} is a list that contains \code{hash} and \code{table}.
\code{hash} is a hash table with keys for example "1-5" (if there are two
state elements, meaning the currently evaluated region lies in the first block
of state element 1 and the fifth block of state element 5) and has a value that
is a data frame with \code{value} set to the value of that block and \code{action}
of the action taken in that block. \code{table} is a list of vectors with length
equal to the number of state elements (2 in this case). It tells which regions in the
domain of the state element correspond to which blocks.
}
