% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribution.R
\name{contribution}
\alias{contribution}
\title{Contribution object}
\usage{
contribution(breakpoints, values1, values2 = rep(0, length(values1)),
  store = F)
}
\arguments{
\item{breakpoints}{vector of numerics (breakpoints, x-direction), must be sorted}

\item{values1}{vector of numerics (downstream value)}

\item{values2}{vector of numerics (contribution)}

\item{store}{boolean on whether or not to store}
}
\description{
Each of these functions are right continuous
}
