% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_helpers.R
\name{mixing_refit_normality}
\alias{mixing_refit_normality}
\title{Mixing function: Refit from Gaussian distribution}
\usage{
mixing_refit_normality(mat, ...)
}
\arguments{
\item{mat}{data matrix}

\item{...}{not used}
}
\value{
numeric matrix
}
\description{
Mixing function: Refit from Gaussian distribution
}
