% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_value.R
\name{MV_maximum_likelihood}
\alias{MV_maximum_likelihood}
\title{Missing value treatment via maximum likelihood}
\usage{
MV_maximum_likelihood(mat, pheno = NA, shrinkage = 0.1, ...)
}
\arguments{
\item{mat}{matrix}

\item{pheno}{data frame}

\item{shrinkage}{amount of regularization to covariance matrix}

\item{...}{not used}
}
\value{
matrix
}
\description{
Missing value treatment via maximum likelihood
}
