% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_selection.R
\name{SS_cook}
\alias{SS_cook}
\title{Sample selection via Cook's distance}
\usage{
SS_cook(dat, pairs = 50, ...)
}
\arguments{
\item{dat}{data object}

\item{pairs}{maximum number of pairs to look at}

\item{...}{not used}
}
\description{
Applies many pairwise (simple) regression and removes points that have
too high of a Cook's distance
}
