% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_selection.R
\name{SS_none}
\alias{SS_none}
\title{Sample selection via all}
\usage{
SS_none(dat, ...)
}
\arguments{
\item{dat}{data object}

\item{...}{not used}
}
\value{
list containing the modified \code{mat} and \code{pheno}
}
\description{
Sample selection via all
}
