% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_selection.R
\name{SS_three_sd}
\alias{SS_three_sd}
\title{Sample selection via three standard deviations}
\usage{
SS_three_sd(mat, pheno = NA, ...)
}
\arguments{
\item{mat}{matrix}

\item{pheno}{data frame}

\item{...}{not used}
}
\value{
list containing the modified \code{mat} and \code{pheno}
}
\description{
Sample selection via three standard deviations
}
